+++ lib/screens/settings/personal_preferences.dart
@@ -529,6 +529,7 @@ class _PersonalPreferencesState extends State<PersonalPreferences> {
       child: Scaffold(
         backgroundColor: Colors.transparent,
         appBar: AppBar(
+          scrolledUnderElevation: 0,
           backgroundColor: darkenColor(appColorNotifier.value, 0.025),
           centerTitle: true,
           toolbarHeight: Responsive.height(context, 100),